Changed all stats to be doubles to allow for easier conversion to p90

+++ Client/DTOs/OverallStatsDto.cs
@@ -5,55 +5,55 @@
         public int MatchesPlayed { get; set; }
 
         #region Goalkeeper Stats
-        public int SavedShotsFromInsideTheBox { get; set; }
-        public int Saves { get; set; }
-        public int Punches { get; set; }
-        public int TotalKeeperSweeper { get; set; }
-        public int AccurateKeeperSweeper { get; set; }
+        public double SavedShotsFromInsideTheBox { get; set; }
+        public double Saves { get; set; }
+        public double Punches { get; set; }
+        public double TotalKeeperSweeper { get; set; }
+        public double AccurateKeeperSweeper { get; set; }
         #endregion
 
         #region Defensive Stats
-        public int ErrorLeadToAShot { get; set; }
-        public int TotalClearance { get; set; }
-        public int AerialLost { get; set; }
-        public int AerialWon { get; set; }
-        public int DuelLost { get; set; }
-        public int DuelWon { get; set; }
-        public int InterceptionWon { get; set; }
-        public int TotalTackle { get; set; }
-        public int OutfielderBlock { get; set; }
+        public double ErrorLeadToAShot { get; set; }
+        public double TotalClearance { get; set; }
+        public double AerialLost { get; set; }
+        public double AerialWon { get; set; }
+        public double DuelLost { get; set; }
+        public double DuelWon { get; set; }
+        public double InterceptionWon { get; set; }
+        public double TotalTackle { get; set; }
+        public double OutfielderBlock { get; set; }
         #endregion
 
         #region Passing Stats
-        public int TotalPass { get; set; }
-        public int GoalAssist { get; set; }
-        public int AccuratePass { get; set; }
-        public int TotalLongBalls { get; set; }
-        public int AccurateLongBalls { get; set; }
-        public int TotalCross { get; set; }
-        public int AccurateCross { get; set; }
-        public int BigChanceCreated { get; set; }
-        public int PossessionLostCtrl { get; set; }
-        public int KeyPass { get; set; }
+        public double TotalPass { get; set; }
+        public double GoalAssist { get; set; }
+        public double AccuratePass { get; set; }
+        public double TotalLongBalls { get; set; }
+        public double AccurateLongBalls { get; set; }
+        public double TotalCross { get; set; }
+        public double AccurateCross { get; set; }
+        public double BigChanceCreated { get; set; }
+        public double PossessionLostCtrl { get; set; }
+        public double KeyPass { get; set; }
         #endregion
 
         #region Shooting Stats
-        public int Goals { get; set; }
-        public int BigChanceMissed { get; set; }
-        public int ShotOffTarget { get; set; }
-        public int OnTargetScoringAttempt { get; set; }
-        public int BlockedScoringAttempt { get; set; }
+        public double Goals { get; set; }
+        public double BigChanceMissed { get; set; }
+        public double ShotOffTarget { get; set; }
+        public double OnTargetScoringAttempt { get; set; }
+        public double BlockedScoringAttempt { get; set; }
         #endregion
 
         #region Misc Stats
-        public int WasFouled { get; set; }
-        public int Fouls { get; set; }
-        public int MinutesPlayed { get; set; }
-        public int Touches { get; set; }
+        public double WasFouled { get; set; }
+        public double Fouls { get; set; }
+        public double MinutesPlayed { get; set; }
+        public double Touches { get; set; }
         public double Rating { get; set; }
-        public int TotalOffside { get; set; }
-        public int Substitute { get; set; }
-        public int Captain { get; set; }
+        public double TotalOffside { get; set; }
+        public double Substitute { get; set; }
+        public double Captain { get; set; }
         #endregion
     }
 }
